+++ src/blob/index/mod.rs
@@ -1,5 +1,3 @@
-use crate::prelude::*;
-
 mod bptree;
 mod core;
 mod header;
