+++ src/filter/ahash/mod.rs
@@ -19,6 +19,8 @@
 mod convert;
 mod fallback_hash;
 mod operations;
+#[cfg(test)]
+mod compatibility_test;
 
 #[doc(hidden)]
 pub const PI: [u64; 4] = [
