+++ tests/tests.rs
@@ -11,7 +11,6 @@ use pearl::{BloomProvider, Builder, Meta, Storage};
 use rand::{seq::SliceRandom, Rng};
 use std::{
     fs,
-    hash::Hasher,
     time::{Duration, Instant},
 };
 use tokio::time::sleep;
@@ -19,22 +18,6 @@ use tokio::time::sleep;
 mod common;
 
 use common::KeyTest;
-
-#[test]
-fn test_hash_algorithm_compat() {
-    test_hash(&(0..10).collect::<Vec<u8>>(), 3604729491498336444);
-    test_hash(&(245..255).collect::<Vec<u8>>(), 4698010058046694585);
-    test_hash(&(63..73).collect::<Vec<u8>>(), 7892047681755360091);
-    test_hash(&(101..111).collect::<Vec<u8>>(), 15822444892006722439);
-}
-
-fn test_hash(data: &[u8], eq_to: u64) {
-    let mut hasher_7 = pearl::ahash::AHasher::new_with_keys(1, 2);
-    hasher_7.write(data);
-    let hash_7 = hasher_7.finish();
-
-    assert_eq!(hash_7, eq_to);
-}
 
 #[tokio::test]
 async fn test_storage_init_new() {
